Modularize shaders
